+++ packages/flame/lib/src/components/sprite_component.dart
@@ -68,6 +68,15 @@ class SpriteComponent extends PositionComponent with HasPaint {
           priority: priority,
         );
 
+  @override
+  @mustCallSuper
+  void onMount() {
+    assert(
+      sprite != null,
+      'You have to set the sprite in either the constructor or in onLoad',
+    );
+  }
+
   @mustCallSuper
   @override
   void render(Canvas canvas) {