+++ packages/flame/CHANGELOG.md
@@ -34,6 +34,7 @@
  - Use `FlameTester` everywhere where it makes sense in the tests
  - Improved `IsometricTileMap`
  - Fix `PositionComponent`'s `flipHorizontallyAroundCenter` and `flipVerticallyAroundCenter`
+ - Initialization of all `PositionComponent`s can be done from `onLoad` instead of the constructor
  - Rename `HasTappableComponents` to `HasTappables`
  - Rename `HasDraggableComponents` to `HasDraggables`
  - Rename `HasHoverableComponents` to `HasHoverableis`